fix(benchmark): render Depth + 2Q gates as integers, not 303.0000
Same category-error fix as the Backends Qubits column. Depth and
two-qubit gate counts are integer-valued; uitable was rendering
them with the default 4-decimal format because the cells were
typed as double. Cast to int32 in both code paths:

  * JsonHelper.benchmarkStrategyToRows — the API success path
    (compare-strategies endpoint response).
  * BenchmarkViewModel local-fallback row builder — the path used
    when the server's transpilation comparison call fails or
    returns no strategies.

Both paths now produce int32 cells, so the Transpilation Strategy
Comparison table reads '303' instead of '303.0000', matching the
Backends Qubits column convention.

+++ src/presentation/viewmodels/BenchmarkViewModel.m
@@ -590,8 +590,11 @@ classdef BenchmarkViewModel < handle
                 comment = sprintf('opt_level=%d (estimated)', lev);
 
                 rows{lev, 1} = name;
-                rows{lev, 2} = depth;
-                rows{lev, 3} = gates;
+                % int32 so uitable renders '100' not '100.0000' — same
+                % convention as the API path in
+                % JsonHelper.benchmarkStrategyToRows.
+                rows{lev, 2} = int32(depth);
+                rows{lev, 3} = int32(gates);
                 rows{lev, 4} = round(fid, 4);
                 rows{lev, 5} = comment;
             end
